refactor(mcp): split server into modules
Extract the monolithic `src/mcp.rs` into `src/mcp/{mod,tools/*}.rs`:

- `mcp/mod.rs` keeps the JSON-RPC scaffolding (request/response wire
  format, dispatch, `Services` shared deps).
- `mcp/tools/mod.rs` re-exports each tool's entry point and hosts
  shared helpers (`resolve_scope`, `apply_budget`).
- One file per MCP tool (`tools/{definitions,deps,diff,edit,files,
  read,search,session}.rs`).

No behavior changes — inline prompts and tool surfaces are intact, so
follow-up MCP prompt / v2-surface work can be reviewed separately.

Forwarded from paulnsorensen/tilth#33.
