+++ src/mcp/mod.rs
@@ -1,0 +1,732 @@
+use std::io::{self, BufRead, Write};
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+use crate::cache::OutlineCache;
+use crate::index::bloom::BloomFilterCache;
+use crate::session::Session;
+use crate::timeout::{self, spawn_with_timeout, SpawnFailure, ThreadTracker};
+
+mod tools;
+
+#[cfg(test)]
+use tools::{parse_file_edit, resolve_scope};
+use tools::{
+    tool_definitions, tool_deps, tool_diff, tool_edit, tool_files, tool_read, tool_search,
+    tool_session,
+};
+
+/// Shared dependencies passed through the request → dispatch pipeline.
+#[derive(Clone)]
+pub(crate) struct Services {
+    cache: Arc<OutlineCache>,
+    session: Arc<Session>,
+    bloom: Arc<BloomFilterCache>,
+    tracker: Arc<ThreadTracker>,
+    edit_mode: bool,
+}
+
+impl Services {
+    pub(crate) fn new(edit_mode: bool) -> Self {
+        Self {
+            cache: Arc::new(OutlineCache::new()),
+            session: Arc::new(Session::new()),
+            bloom: Arc::new(BloomFilterCache::new()),
+            tracker: Arc::new(ThreadTracker::new()),
+            edit_mode,
+        }
+    }
+
+    pub(crate) fn cache(&self) -> &OutlineCache {
+        &self.cache
+    }
+
+    pub(crate) fn session(&self) -> &Session {
+        &self.session
+    }
+
+    pub(crate) fn bloom(&self) -> &Arc<BloomFilterCache> {
+        &self.bloom
+    }
+
+    pub(crate) fn tracker(&self) -> &Arc<ThreadTracker> {
+        &self.tracker
+    }
+
+    pub(crate) fn edit_mode(&self) -> bool {
+        self.edit_mode
+    }
+}
+
+// Sent to the LLM via the MCP `instructions` field during initialization.
+// Keeps the strategic guidance from AGENTS.md available to any host.
+const SERVER_INSTRUCTIONS: &str = "\
+tilth — code intelligence MCP server. Replaces grep, cat, find, ls with AST-aware equivalents.\n\
+\n\
+To explore code, always search first. tilth_search finds definitions, usages, and file locations in one call.\n\
+Usage: tilth_search(query: \"handleRequest\").\n\
+tilth_files is ONLY for listing directory contents when you have no symbol or text to search for.\n\
+DO NOT use Read if content is already shown in expanded search results.\n\
+DO NOT use Grep, Read, or Glob. Always use the better tools tilth_search (grep), tilth_read (read), tilth_files (glob).\n\
+\n\
+tilth_search: Search code — finds definitions, usages, and text. Replaces grep/rg for all code search.\n\
+  For multi-symbol lookup, separate each with a comma \"symbol1,symbol2\" (max 5).\n\
+  kind: \"symbol\" (default) | \"content\" (strings/comments) | \"callers\" (call sites)\n\
+  expand (default 2): inline full source for top matches.\n\
+  context: path to file being edited — boosts nearby results.\n\
+  glob: file pattern filter — \"*.rs\" (whitelist), \"!*.test.ts\" (exclude).\n\
+  Output per match:\n\
+    ## <path>:<start>-<end> [definition|usage|impl]\n\
+    <outline context>\n\
+    <expanded source block>\n\
+    ── calls ──\n\
+      <name>  <path>:<start>-<end>  <signature>\n\
+    ── siblings ──\n\
+      <name>  <path>:<start>-<end>  <signature>\n\
+  Re-expanding a previously shown definition returns [shown earlier].\n\
+\n\
+tilth_read: Read file content with smart outlining. Replaces cat/head/tail.\n\
+  Small files → full content. Large files → structural outline.\n\
+  section: \"<start>-<end>\" or \"<heading text>\"\n\
+  sections: array of ranges/headings — multiple slices from the same file in one call.\n\
+  paths: read multiple files in one call.\n\
+  Output:\n\
+    <line_number> │ <content>                  ← full/section mode\n\
+    [<start>-<end>]  <symbol name>             ← outline mode\n\
+\n\
+tilth_files: Find files by glob pattern. Replaces find, ls, pwd, and the host Glob tool.\n\
+  patterns: run multiple globs in one call (e.g. patterns: [\"*.rs\", \"*.toml\"]).\n\
+  Output: <path>  (~<token_count> tokens).\n\
+\n\
+tilth_deps: Blast-radius check — what imports this file and what it imports.\n\
+  Use ONLY before renaming, removing, or changing an export's signature.\n\
+\n\
+tilth_diff: Structural diff — shows what changed at function level. Replaces Bash(git diff).\n\
+  Usage: tilth_diff(source: \"HEAD~1\") for last commit. No args = uncommitted changes.\n\
+  scope: \"file.rs\" or \"file.rs:fn_name\". log: \"HEAD~5..HEAD\" for per-commit summaries.\n\
+  search: filter to lines matching a term. blast: true to show callers of changed signatures.\n\
+  Output: [+] added, [-] deleted, [~] body changed, [~:sig] signature changed.\n\
+  DO NOT use Bash(git diff) or Bash(git log --patch). Use tilth_diff instead.\n\
+\n\
+To search code, use tilth_search instead of Grep or Bash(grep/rg).\n\
+To read files, use tilth_read instead of Read or Bash(cat).\n\
+To find files, use tilth_files instead of Glob or Bash(find/ls).\n\
+To check what changed, use tilth_diff instead of Bash(git diff/git log).\n\
+DO NOT re-read files already shown in expanded search results.";
+
+const EDIT_MODE_EXTRA: &str = "\n\
+\n\
+tilth_edit: Batch edit one or more files using hash-anchored lines. Replaces the host Edit tool.\n\
+  ALWAYS group edits to multiple files into ONE tilth_edit call (max 20 files). Never call tilth_edit twice in a row.\n\
+  tilth_read → copy anchors (<line>:<hash>) (BOTH line and hash required) → pass to tilth_edit.\n\
+  tilth_search does NOT provide hashes — you MUST tilth_read the file or section first.\n\
+  Shape: {\"files\": [{\"path\": \"a.rs\", \"edits\": [...]}, {\"path\": \"b.rs\", \"edits\": [...]}]}\n\
+  Single file: {\"files\": [{\"path\": \"a.rs\", \"edits\": [{\"start\": \"<line>:<hash>\", \"content\": \"<new code>\"}]}]}\n\
+  Edit forms inside `edits`:\n\
+    Single line: {\"start\": \"<line>:<hash>\", \"content\": \"<new code>\"}\n\
+    Range:       {\"start\": \"<line>:<hash>\", \"end\": \"<line>:<hash>\", \"content\": \"...\"}\n\
+    Delete:      {\"start\": \"<line>:<hash>\", \"content\": \"\"}\n\
+  Per-file results: each file is processed independently. A hash mismatch on one file does NOT block the others.\n\
+  isError is false whenever ≥1 file succeeded — always scan the per-file `## <path>` sections for failures rather than trusting the top-level status.\n\
+  Hash mismatch → file changed, re-read THAT file and retry it (other files in the batch already applied).\n\
+  A parse error on one edit invalidates ALL edits for that file (none applied); retry the whole file's edits after fixing the malformed entry.\n\
+  Each file path may appear at most once per call — group all edits for a file under its single entry.\n\
+  Large files: tilth_read shows outline — use section to get hashlined content.\n\
+  Pass diff: true to see a compact before/after diff per file.\n\
+  After editing a function signature, tilth_edit shows callers that may need updating.\n\
+DO NOT use the host Edit tool. Use tilth_edit for all edits.";
+
+/// MCP server over stdio. When `edit_mode` is true, exposes `tilth_edit` and
+/// switches `tilth_read` to hashline output format.
+///
+/// `scope` overrides the default search root. When provided, tilth chdir's to it
+/// at startup so all tools, git commands, and searches use the correct project root.
+/// This fixes MCP hosts that launch tilth with cwd=/ (e.g., Codex).
+pub fn run(edit_mode: bool, scope: Option<&Path>) -> io::Result<()> {
+    let scope_is_explicit = scope.is_some();
+
+    // Resolve the project root and chdir to it.
+    // Priority: explicit --scope > MCP roots (handled later) > package_root(cwd) > cwd
+    if let Some(s) = scope {
+        if s.is_dir() {
+            let _ = std::env::set_current_dir(s);
+        }
+    } else {
+        let cwd = std::env::current_dir().unwrap_or_default();
+        if let Some(root) = crate::lang::package_root(&cwd) {
+            let _ = std::env::set_current_dir(root);
+        }
+    }
+
+    let services = Services::new(edit_mode);
+    let stdin = io::stdin();
+    let stdout = io::stdout();
+    let mut stdout = stdout.lock();
+
+    // Track pending roots/list request (for MCP roots protocol)
+    let mut pending_roots_id: Option<Value> = None;
+
+    for line in stdin.lock().lines() {
+        let line = line?;
+        if line.is_empty() {
+            continue;
+        }
+
+        // Parse as generic JSON first — could be a request, notification, or response
+        let msg: Value = match serde_json::from_str(&line) {
+            Ok(v) => v,
+            Err(e) => {
+                write_error(&mut stdout, None, -32700, &format!("parse error: {e}"))?;
+                continue;
+            }
+        };
+
+        // Check if this is a response to our roots/list request
+        if let Some(ref roots_id) = pending_roots_id {
+            if msg.get("id") == Some(roots_id) {
+                pending_roots_id = None;
+                // Only apply roots on success and if --scope was NOT explicitly provided
+                if !scope_is_explicit {
+                    if let Some(root_path) = extract_root_from_response(&msg) {
+                        let _ = std::env::set_current_dir(&root_path);
+                    }
+                }
+                continue;
+            }
+        }
+
+        // Must have "method" to be a request or notification
+        let method = match msg.get("method").and_then(Value::as_str) {
+            Some(m) => m.to_string(),
+            None => continue, // Not a request — skip (could be an unexpected response)
+        };
+
+        let id = msg.get("id").cloned();
+        if id.is_none() {
+            // Notifications have no id — silently drop them per JSON-RPC spec
+            continue;
+        }
+
+        // Parse params
+        let params = msg.get("params").cloned().unwrap_or(Value::Null);
+
+        let req = JsonRpcRequest {
+            _jsonrpc: "2.0".to_string(),
+            id,
+            method: method.clone(),
+            params,
+        };
+
+        let response = handle_request(&req, &services);
+        serde_json::to_writer(&mut stdout, &response)?;
+        stdout.write_all(b"\n")?;
+        stdout.flush()?;
+
+        // After initialize response: send roots/list if client supports it
+        // and we don't already have an explicit --scope
+        if method == "initialize" && !scope_is_explicit && pending_roots_id.is_none() {
+            let client_caps = req.params.get("capabilities").unwrap_or(&Value::Null);
+            if client_caps.get("roots").is_some() {
+                let roots_id = Value::String("tilth_roots_1".to_string());
+                let roots_req = serde_json::json!({
+                    "jsonrpc": "2.0",
+                    "id": roots_id,
+                    "method": "roots/list"
+                });
+                serde_json::to_writer(&mut stdout, &roots_req)?;
+                stdout.write_all(b"\n")?;
+                stdout.flush()?;
+                pending_roots_id = Some(roots_id);
+            }
+        }
+    }
+
+    Ok(())
+}
+
+/// Extract the first root directory path from a roots/list response.
+/// Parses `file://` URIs and returns the path, or None if no valid roots.
+fn extract_root_from_response(msg: &Value) -> Option<PathBuf> {
+    let roots = msg.get("result")?.get("roots")?.as_array()?;
+    for root in roots {
+        let uri = root.get("uri")?.as_str()?;
+        let raw_path = uri.strip_prefix("file://").unwrap_or(uri);
+        // On invalid UTF-8 in a percent-encoded path, fall back to the
+        // original input rather than substituting U+FFFD replacements.
+        let decoded = percent_encoding::percent_decode_str(raw_path)
+            .decode_utf8()
+            .map_or_else(|_| raw_path.to_string(), std::borrow::Cow::into_owned);
+        let path = PathBuf::from(decoded);
+        if path.is_dir() {
+            return Some(path);
+        }
+    }
+    None
+}
+
+#[derive(Deserialize)]
+struct JsonRpcRequest {
+    #[serde(rename = "jsonrpc")]
+    _jsonrpc: String,
+    id: Option<Value>,
+    method: String,
+    #[serde(default)]
+    params: Value,
+}
+
+#[derive(Serialize)]
+struct JsonRpcResponse {
+    jsonrpc: &'static str,
+    id: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    result: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    error: Option<JsonRpcError>,
+}
+
+#[derive(Serialize)]
+struct JsonRpcError {
+    code: i32,
+    message: String,
+}
+
+fn handle_request(req: &JsonRpcRequest, services: &Services) -> JsonRpcResponse {
+    let edit_mode = services.edit_mode();
+    match req.method.as_str() {
+        "initialize" => {
+            let overview = if std::env::var("TILTH_NO_OVERVIEW").is_ok() {
+                String::new()
+            } else {
+                let cwd = std::env::current_dir().unwrap_or_default();
+                crate::overview::fingerprint(&cwd)
+            };
+            let instructions = if edit_mode {
+                if overview.is_empty() {
+                    format!("{SERVER_INSTRUCTIONS}{EDIT_MODE_EXTRA}")
+                } else {
+                    format!("{overview}\n\n{SERVER_INSTRUCTIONS}{EDIT_MODE_EXTRA}")
+                }
+            } else if overview.is_empty() {
+                SERVER_INSTRUCTIONS.to_string()
+            } else {
+                format!("{overview}\n\n{SERVER_INSTRUCTIONS}")
+            };
+            JsonRpcResponse {
+                jsonrpc: "2.0",
+                id: req.id.clone(),
+                result: Some(serde_json::json!({
+                    "protocolVersion": "2024-11-05",
+                    "capabilities": {
+                        "tools": {}
+                    },
+                    "serverInfo": {
+                        "name": "tilth",
+                        "version": env!("CARGO_PKG_VERSION")
+                    },
+                    "instructions": instructions
+                })),
+                error: None,
+            }
+        }
+
+        "tools/list" => JsonRpcResponse {
+            jsonrpc: "2.0",
+            id: req.id.clone(),
+            result: Some(serde_json::json!({
+                "tools": tool_definitions(edit_mode)
+            })),
+            error: None,
+        },
+
+        "tools/call" => handle_tool_call(req, services),
+
+        "ping" => JsonRpcResponse {
+            jsonrpc: "2.0",
+            id: req.id.clone(),
+            result: Some(serde_json::json!({})),
+            error: None,
+        },
+
+        _ => JsonRpcResponse {
+            jsonrpc: "2.0",
+            id: req.id.clone(),
+            result: None,
+            error: Some(JsonRpcError {
+                code: -32601,
+                message: format!("method not found: {}", req.method),
+            }),
+        },
+    }
+}
+
+/// Execute a tool by name with the given arguments. Returns formatted output or error string.
+/// No classifier involved — the caller specifies the tool explicitly.
+pub(crate) fn dispatch_tool(
+    tool: &str,
+    args: &Value,
+    services: &Services,
+) -> Result<String, String> {
+    let edit_mode = services.edit_mode();
+    match tool {
+        "tilth_read" => tool_read(args, services.cache(), services.session(), edit_mode),
+        "tilth_search" => tool_search(args, services.cache(), services.session(), services.bloom()),
+        "tilth_files" => tool_files(args),
+        "tilth_deps" => tool_deps(args, services.bloom()),
+        "tilth_diff" => tool_diff(args),
+        "tilth_session" => tool_session(args, services.session()),
+        "tilth_edit" if edit_mode => tool_edit(args, services.session(), services.bloom()),
+        _ => Err(format!("unknown tool: {tool}")),
+    }
+}
+
+// ---------------------------------------------------------------------------
+// MCP tool call handler
+// ---------------------------------------------------------------------------
+
+fn handle_tool_call(req: &JsonRpcRequest, services: &Services) -> JsonRpcResponse {
+    let params = &req.params;
+    let tool_name = params.get("name").and_then(|v| v.as_str()).unwrap_or("");
+    let args = params.get("arguments").unwrap_or(&Value::Null);
+
+    let result = if services.tracker().is_at_cap() {
+        Err(
+            "server busy: too many prior operations still running after timeout. \
+             Wait or set TILTH_TIMEOUT=<seconds> higher."
+                .into(),
+        )
+    } else {
+        run_tool_with_timeout(services, tool_name, args, timeout::request_timeout())
+    };
+
+    build_tool_response(req.id.clone(), result)
+}
+
+fn run_tool_with_timeout(
+    services: &Services,
+    tool_name: &str,
+    args: &Value,
+    timeout: std::time::Duration,
+) -> Result<String, String> {
+    let services_worker = services.clone();
+    let tool_name_owned = tool_name.to_string();
+    let args_owned = args.clone();
+
+    let outcome = spawn_with_timeout(services.tracker(), timeout, move || {
+        dispatch_tool(&tool_name_owned, &args_owned, &services_worker)
+    });
+
+    match outcome {
+        Ok(inner) => inner,
+        Err(SpawnFailure::Timeout) => Err(format!(
+            "tool timed out after {}s — the operation took too long. \
+             Try: reduce scope, use section instead of full, or set \
+             TILTH_TIMEOUT=<seconds> to increase the limit.",
+            timeout.as_secs()
+        )),
+        Err(SpawnFailure::Panic) => Err("tool panicked during execution".into()),
+    }
+}
+
+fn build_tool_response(id: Option<Value>, result: Result<String, String>) -> JsonRpcResponse {
+    let (text, is_error) = match result {
+        Ok(output) => (output, false),
+        Err(e) => (e, true),
+    };
+    let mut payload = serde_json::json!({
+        "content": [{ "type": "text", "text": text }]
+    });
+    if is_error {
+        payload["isError"] = Value::Bool(true);
+    }
+    JsonRpcResponse {
+        jsonrpc: "2.0",
+        id,
+        result: Some(payload),
+        error: None,
+    }
+}
+
+fn write_error(w: &mut impl Write, id: Option<Value>, code: i32, msg: &str) -> io::Result<()> {
+    let resp = JsonRpcResponse {
+        jsonrpc: "2.0",
+        id,
+        result: None,
+        error: Some(JsonRpcError {
+            code,
+            message: msg.into(),
+        }),
+    };
+    serde_json::to_writer(&mut *w, &resp)?;
+    w.write_all(b"\n")?;
+    w.flush()
+}
+
+// ---------------------------------------------------------------------------
+// Tests
+// ---------------------------------------------------------------------------
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    // -- extract_root_from_response -------------------------------------------
+
+    #[test]
+    fn extract_root_valid_file_uri() {
+        // Claude Code sends: {"result":{"roots":[{"uri":"file:///Users/x/project"}]}}
+        let tmp = tempfile::tempdir().unwrap();
+        let uri = format!("file://{}", tmp.path().display());
+        let msg = serde_json::json!({
+            "result": { "roots": [{ "uri": uri }] }
+        });
+        let path = extract_root_from_response(&msg);
+        assert_eq!(path, Some(tmp.path().to_path_buf()));
+    }
+
+    #[test]
+    fn extract_root_percent_encoded_uri() {
+        let tmp = tempfile::tempdir().unwrap();
+        let space_dir = tmp.path().join("my project");
+        std::fs::create_dir(&space_dir).unwrap();
+        let encoded =
+            format!("file://{}", tmp.path().display()).replace(' ', "%20") + "/my%20project";
+        let msg = serde_json::json!({
+            "result": { "roots": [{ "uri": encoded }] }
+        });
+        let path = extract_root_from_response(&msg);
+        assert_eq!(path, Some(space_dir));
+    }
+
+    #[test]
+    fn extract_root_empty_roots() {
+        // Codex sends: {"result":{"roots":[]}}
+        let msg = serde_json::json!({
+            "result": { "roots": [] }
+        });
+        assert_eq!(extract_root_from_response(&msg), None);
+    }
+
+    #[test]
+    fn extract_root_nonexistent_path() {
+        let msg = serde_json::json!({
+            "result": { "roots": [{ "uri": "file:///nonexistent/path/that/does/not/exist" }] }
+        });
+        assert_eq!(extract_root_from_response(&msg), None);
+    }
+
+    #[test]
+    fn extract_root_no_result() {
+        let msg = serde_json::json!({"error": {"code": -1, "message": "nope"}});
+        assert_eq!(extract_root_from_response(&msg), None);
+    }
+
+    #[test]
+    fn extract_root_multiple_roots_takes_first_valid() {
+        let tmp = tempfile::tempdir().unwrap();
+        let uri = format!("file://{}", tmp.path().display());
+        let msg = serde_json::json!({
+            "result": { "roots": [
+                { "uri": "file:///nonexistent" },
+                { "uri": uri },
+            ]}
+        });
+        // First root is invalid, second is valid — should return second
+        let path = extract_root_from_response(&msg);
+        assert_eq!(path, Some(tmp.path().to_path_buf()));
+    }
+
+    // -- resolve_scope --------------------------------------------------------
+
+    #[test]
+    fn resolve_scope_explicit_arg() {
+        let tmp = tempfile::tempdir().unwrap();
+        let args = serde_json::json!({ "scope": tmp.path().to_str().unwrap() });
+        let (scope, warning) = resolve_scope(&args);
+        assert_eq!(scope, tmp.path().canonicalize().unwrap());
+        assert!(warning.is_none());
+    }
+
+    #[test]
+    fn resolve_scope_no_arg_uses_cwd() {
+        let args = serde_json::json!({});
+        let (scope, warning) = resolve_scope(&args);
+        // With no arg, defaults to "." which is cwd
+        let cwd = std::env::current_dir().unwrap();
+        // The function returns "." when resolved == cwd
+        assert!(scope == PathBuf::from(".") || scope == cwd);
+        assert!(warning.is_none());
+    }
+
+    #[test]
+    fn resolve_scope_invalid_dir_warns() {
+        let args = serde_json::json!({ "scope": "/nonexistent/directory/zzz" });
+        let (scope, warning) = resolve_scope(&args);
+        assert_eq!(scope, PathBuf::from("."));
+        assert!(warning.is_some());
+        assert!(warning.unwrap().contains("not a valid directory"));
+    }
+
+    // -- Issue #37 reproduction: cwd=/ with --scope fixes it ------------------
+
+    #[test]
+    fn scope_flag_overrides_bad_cwd() {
+        // Reproduce issue #37: MCP host launches tilth with cwd=/
+        // The --scope flag should override this.
+        let project = tempfile::tempdir().unwrap();
+        let project_path = project.path();
+
+        // Create a manifest so package_root can find it
+        std::fs::write(
+            project_path.join("Cargo.toml"),
+            "[package]\nname = \"test\"",
+        )
+        .unwrap();
+        std::fs::create_dir(project_path.join("src")).unwrap();
+        std::fs::write(project_path.join("src/main.rs"), "fn main() {}").unwrap();
+
+        // Save current cwd
+        let orig_cwd = std::env::current_dir().unwrap();
+
+        // Simulate Codex: cwd=/
+        std::env::set_current_dir("/").unwrap();
+
+        // Without --scope: resolve_scope returns "." which is /
+        let args = serde_json::json!({});
+        let (scope, _) = resolve_scope(&args);
+        assert_eq!(
+            scope,
+            PathBuf::from("."),
+            "Without --scope, should return . (which is /)"
+        );
+
+        // With --scope pointing to project: set_current_dir should fix everything
+        let _ = std::env::set_current_dir(project_path);
+        let args = serde_json::json!({});
+        let (scope, _) = resolve_scope(&args);
+        assert_eq!(
+            scope,
+            PathBuf::from("."),
+            "After chdir to project, . should resolve correctly"
+        );
+
+        // Verify tilth_files would search in the project, not /
+        let cwd = std::env::current_dir().unwrap();
+        assert_eq!(cwd, project_path.canonicalize().unwrap());
+
+        // Restore
+        std::env::set_current_dir(orig_cwd).unwrap();
+    }
+
+    // -- tool_files multi-pattern --------------------------------------------
+
+    /// Build a small scratch project with .rs and .toml files, switch cwd to
+    /// it, and return the tempdir guard so the caller controls cleanup.
+    fn scratch_project() -> tempfile::TempDir {
+        let project = tempfile::tempdir().unwrap();
+        let p = project.path();
+        std::fs::write(p.join("Cargo.toml"), "[package]\nname = \"t\"").unwrap();
+        std::fs::create_dir(p.join("src")).unwrap();
+        std::fs::write(p.join("src/main.rs"), "fn main() {}").unwrap();
+        std::fs::write(p.join("src/lib.rs"), "pub fn x() {}").unwrap();
+        project
+    }
+
+    #[test]
+    fn tool_files_patterns_emits_one_block_per_pattern() {
+        let project = scratch_project();
+        let args = serde_json::json!({
+            "patterns": ["*.rs", "*.toml"],
+            "scope": project.path().to_str().unwrap(),
+        });
+        let out = tool_files(&args).expect("tool_files should succeed");
+        // Two `# Glob:` headers — one per pattern.
+        let header_count = out.matches("# Glob:").count();
+        assert_eq!(header_count, 2, "expected 2 Glob headers, got: {out}");
+        assert!(out.contains("\"*.rs\""), "missing rs header in: {out}");
+        assert!(out.contains("\"*.toml\""), "missing toml header in: {out}");
+        // Files from both patterns appear in the combined output.
+        assert!(out.contains("main.rs"), "missing main.rs in: {out}");
+        assert!(out.contains("Cargo.toml"), "missing Cargo.toml in: {out}");
+    }
+
+    #[test]
+    fn tool_files_pattern_and_patterns_mutually_exclusive() {
+        let args = serde_json::json!({
+            "pattern": "*.rs",
+            "patterns": ["*.rs"],
+        });
+        let err = tool_files(&args).expect_err("expected mutual-exclusion error");
+        assert!(err.contains("either pattern"), "unexpected error: {err}");
+    }
+
+    #[test]
+    fn tool_files_empty_patterns_errors() {
+        let args = serde_json::json!({ "patterns": [] });
+        let err = tool_files(&args).expect_err("expected empty-patterns error");
+        assert!(err.contains("at least one"), "unexpected error: {err}");
+    }
+
+    #[test]
+    fn tool_files_patterns_capped_at_20() {
+        let twenty_one: Vec<&str> = vec!["*.rs"; 21];
+        let args = serde_json::json!({ "patterns": twenty_one });
+        let err = tool_files(&args).expect_err("expected cap error");
+        assert!(err.contains("limited to 20"), "unexpected error: {err}");
+    }
+
+    #[test]
+    fn tool_files_missing_pattern_and_patterns_errors() {
+        let args = serde_json::json!({});
+        let err = tool_files(&args).expect_err("expected missing-pattern error");
+        assert!(
+            err.contains("missing required parameter"),
+            "unexpected error: {err}"
+        );
+    }
+
+    // -- parse_file_edit edge cases -------------------------------------------
+
+    #[test]
+    fn parse_file_edit_rejects_empty_edits_array() {
+        // Schema says minItems: 1, but schema validation is advisory — enforce
+        // at runtime so a client that bypasses the schema can't silently get
+        // a no-op success.
+        let val = serde_json::json!({ "path": "noop.txt", "edits": [] });
+        let task = parse_file_edit(0, &val);
+        match task {
+            crate::edit::FileEditTask::ParseError { label, msg } => {
+                assert_eq!(label, "noop.txt");
+                assert!(msg.contains("empty"), "unexpected msg: {msg}");
+            }
+            crate::edit::FileEditTask::Ready { .. } => {
+                panic!("empty edits array should produce a ParseError, not Ready");
+            }
+        }
+    }
+
+    // -- package_root fallback from subdirectory ------------------------------
+
+    #[test]
+    fn package_root_finds_project_from_subdirectory() {
+        let project = tempfile::tempdir().unwrap();
+        let project_path = project.path();
+        std::fs::write(
+            project_path.join("Cargo.toml"),
+            "[package]\nname = \"test\"",
+        )
+        .unwrap();
+        let subdir = project_path.join("src").join("deep").join("nested");
+        std::fs::create_dir_all(&subdir).unwrap();
+
+        // package_root from the nested subdir should find the project root
+        let root = crate::lang::package_root(&subdir);
+        assert!(root.is_some(), "package_root should find the project");
+        // Compare canonicalized paths to handle macOS /var -> /private/var symlinks
+        let root_canon = root.unwrap().canonicalize().unwrap();
+        let expected_canon = project_path.canonicalize().unwrap();
+        assert_eq!(root_canon, expected_canon);
+    }
+}
