+++ src/mcp/tools/diff.rs
@@ -1,0 +1,17 @@
+use serde_json::Value;
+
+pub(crate) fn tool_diff(args: &Value) -> Result<String, String> {
+    let source = args.get("source").and_then(|v| v.as_str());
+    let scope = args.get("scope").and_then(|v| v.as_str());
+    let a = args.get("a").and_then(|v| v.as_str());
+    let b = args.get("b").and_then(|v| v.as_str());
+    let patch = args.get("patch").and_then(|v| v.as_str());
+    let log = args.get("log").and_then(|v| v.as_str());
+    let search = args.get("search").and_then(|v| v.as_str());
+    let blast = args.get("blast").and_then(Value::as_bool).unwrap_or(false);
+    let expand = args.get("expand").and_then(Value::as_u64).unwrap_or(0) as usize;
+    let budget = args.get("budget").and_then(Value::as_u64);
+
+    let diff_source = crate::diff::resolve_source(source, a, b, patch, log)?;
+    crate::diff::diff(&diff_source, scope, search, blast, expand, budget)
+}
