+++ src/mcp/tools/edit.rs
@@ -1,0 +1,126 @@
+use std::path::PathBuf;
+use std::sync::Arc;
+
+use serde_json::Value;
+
+use crate::index::bloom::BloomFilterCache;
+use crate::session::Session;
+
+/// Parse one `files[]` entry. Parse errors are deferred onto the task so a
+/// malformed entry surfaces as a per-file failure instead of aborting the
+/// whole batch.
+pub(crate) fn parse_file_edit(index: usize, val: &Value) -> crate::edit::FileEditTask {
+    use crate::edit::FileEditTask;
+
+    let Some(path_str) = val.get("path").and_then(|v| v.as_str()) else {
+        return FileEditTask::ParseError {
+            label: format!("files[{index}]"),
+            msg: "missing 'path'".into(),
+        };
+    };
+    let Some(edits_val) = val.get("edits").and_then(|v| v.as_array()) else {
+        return FileEditTask::ParseError {
+            label: path_str.to_string(),
+            msg: "missing 'edits' array".into(),
+        };
+    };
+
+    if edits_val.is_empty() {
+        return FileEditTask::ParseError {
+            label: path_str.to_string(),
+            msg: "'edits' array is empty — omit this file or add at least one edit".into(),
+        };
+    }
+
+    let mut edits = Vec::with_capacity(edits_val.len());
+    for (i, e) in edits_val.iter().enumerate() {
+        match parse_edit_entry(i, e) {
+            Ok(edit) => edits.push(edit),
+            Err(msg) => {
+                return FileEditTask::ParseError {
+                    label: path_str.to_string(),
+                    msg,
+                };
+            }
+        }
+    }
+
+    FileEditTask::Ready {
+        path: PathBuf::from(path_str),
+        edits,
+    }
+}
+
+/// Parse a single `edits[]` entry. Errors carry the edit index so the LLM
+/// can fix exactly the right entry instead of guessing.
+fn parse_edit_entry(i: usize, e: &Value) -> Result<crate::edit::Edit, String> {
+    let start_str = e
+        .get("start")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| format!("edit[{i}]: missing 'start'"))?;
+    let (start_line, start_hash) = crate::format::parse_anchor(start_str)
+        .ok_or_else(|| format!("edit[{i}]: invalid start anchor '{start_str}'"))?;
+    let (end_line, end_hash) = match e.get("end").and_then(|v| v.as_str()) {
+        Some(end_str) => crate::format::parse_anchor(end_str)
+            .ok_or_else(|| format!("edit[{i}]: invalid end anchor '{end_str}'"))?,
+        None => (start_line, start_hash),
+    };
+    let content = e
+        .get("content")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| format!("edit[{i}]: missing 'content'"))?;
+    Ok(crate::edit::Edit {
+        start_line,
+        start_hash,
+        end_line,
+        end_hash,
+        content: content.to_string(),
+    })
+}
+
+pub(crate) fn tool_edit(
+    args: &Value,
+    session: &Session,
+    bloom: &Arc<BloomFilterCache>,
+) -> Result<String, String> {
+    let files_val = args
+        .get("files")
+        .and_then(|v| v.as_array())
+        .ok_or("missing required parameter: files (array of {path, edits})")?;
+
+    if files_val.is_empty() {
+        return Err("files array is empty".into());
+    }
+    if files_val.len() > 20 {
+        return Err(format!(
+            "batch edit limited to 20 files (got {})",
+            files_val.len()
+        ));
+    }
+
+    let show_diff = args
+        .get("diff")
+        .and_then(serde_json::Value::as_bool)
+        .unwrap_or(false);
+
+    let tasks: Vec<crate::edit::FileEditTask> = files_val
+        .iter()
+        .enumerate()
+        .map(|(i, v)| parse_file_edit(i, v))
+        .collect();
+
+    // Fast-fail on duplicates before touching session state. apply_batch
+    // re-runs the same check as an encapsulation guarantee for any future
+    // caller that bypasses this wire layer.
+    if let Some(msg) = crate::edit::detect_duplicate_paths(&tasks) {
+        return Err(msg);
+    }
+
+    for task in &tasks {
+        if let crate::edit::FileEditTask::Ready { path, .. } = task {
+            session.record_read(path);
+        }
+    }
+
+    crate::edit::apply_batch(tasks, bloom, show_diff)
+}
