+++ src/mcp/tools/session.rs
@@ -1,0 +1,17 @@
+use serde_json::Value;
+
+use crate::session::Session;
+
+pub(crate) fn tool_session(args: &Value, session: &Session) -> Result<String, String> {
+    let action = args
+        .get("action")
+        .and_then(|v| v.as_str())
+        .unwrap_or("summary");
+    match action {
+        "reset" => {
+            session.reset();
+            Ok("Session reset.".to_string())
+        }
+        _ => Ok(session.summary()),
+    }
+}
